+++ src/resdk/resources/data.py
@@ -3,7 +3,7 @@
 import json
 import logging
 import os
-from typing import Optional
+from typing import TYPE_CHECKING, Any, Optional
 from urllib.parse import urljoin
 
 from resdk.constants import CHUNK_SIZE
@@ -11,11 +11,20 @@ from resdk.constants import CHUNK_SIZE
 from ..utils.decorators import assert_object_exists
 from .background_task import BackgroundTask
 from .base import BaseResolweResource
-from .collection import Collection
-from .descriptor import DescriptorSchema
-from .process import Process
-from .sample import Sample
-from .utils import flatten_field, parse_resolwe_datetime
+from .fields import (
+    BooleanField,
+    DataSource,
+    DateTimeField,
+    DictField,
+    DictResourceField,
+    FieldAccessType,
+    IntegerField,
+    StringField,
+)
+from .utils import flatten_field
+
+if TYPE_CHECKING:
+    from resdk.resolwe import Resolwe
 
 
 class Data(BaseResolweResource):
@@ -30,85 +39,56 @@ class Data(BaseResolweResource):
     endpoint = "data"
     full_search_paramater = "text"
 
-    READ_ONLY_FIELDS = BaseResolweResource.READ_ONLY_FIELDS + (
-        "checksum",
-        "descriptor_dirty",
-        "duplicated",
-        "process_cores",
-        "process_error",
-        "process_info",
-        "process_memory",
-        "process_progress",
-        "process_rc",
-        "process_warning",
-        "output",
-        "scheduled",
-        "size",
-        "status",
-    )
-    UPDATE_PROTECTED_FIELDS = BaseResolweResource.UPDATE_PROTECTED_FIELDS + (
-        "input",
-        "process",
-    )
-    WRITABLE_FIELDS = BaseResolweResource.WRITABLE_FIELDS + (
-        "collection",
-        "descriptor",
-        "descriptor_schema",
-        "process_resources",
-        "sample",
-        "tags",
-    )
+    checksum = StringField(access_type=FieldAccessType.READ_ONLY)
+    descriptor_dirty = BooleanField(access_type=FieldAccessType.READ_ONLY)
+    duplicated = DateTimeField(access_type=FieldAccessType.READ_ONLY)
+    process_cores = IntegerField(access_type=FieldAccessType.READ_ONLY)
+    process_error = StringField(access_type=FieldAccessType.READ_ONLY, many=True)
+    process_info = StringField(access_type=FieldAccessType.READ_ONLY, many=True)
+    process_memory = IntegerField(access_type=FieldAccessType.READ_ONLY)
+    process_progress = IntegerField(access_type=FieldAccessType.READ_ONLY)
+    process_rc = IntegerField(access_type=FieldAccessType.READ_ONLY)
+    process_warning = StringField(access_type=FieldAccessType.READ_ONLY, many=True)
+    output = DictField(access_type=FieldAccessType.READ_ONLY)
+    scheduled = DateTimeField(access_type=FieldAccessType.READ_ONLY)
+    size = IntegerField(access_type=FieldAccessType.READ_ONLY)
+    status = StringField(access_type=FieldAccessType.READ_ONLY)
 
-    def __init__(self, resolwe, **model_data):
+    input = DictField(access_type=FieldAccessType.UPDATE_PROTECTED)
+    process = DictResourceField(
+        resource_class_name="Process",
+        property_name="slug",
+        access_type=FieldAccessType.UPDATE_PROTECTED,
+    )
+    collection = DictResourceField(
+        resource_class_name="Collection", access_type=FieldAccessType.WRITABLE
+    )
+    descriptor = DictField(access_type=FieldAccessType.WRITABLE)
+    descriptor_schema = DictResourceField(
+        resource_class_name="DescriptorSchema",
+        property_name="slug",
+        access_type=FieldAccessType.WRITABLE,
+    )
+    process_resources = DictField(access_type=FieldAccessType.WRITABLE)
+
+    sample = DictResourceField(
+        resource_class_name="Sample",
+        server_field_name="entity",
+        access_type=FieldAccessType.WRITABLE,
+    )
+    tags = StringField(access_type=FieldAccessType.WRITABLE, many=True)
+    started = DateTimeField(access_type=FieldAccessType.READ_ONLY)
+    finished = DateTimeField(access_type=FieldAccessType.READ_ONLY)
+
+    def __init__(self, resolwe: "Resolwe", **model_data: dict):
         """Initialize attributes."""
         self.logger = logging.getLogger(__name__)
-
-        #: ``Collection``s that contains ``Data``
-        self._collection = None
-        #: ``DescriptorSchema`` of ``Data`` object
-        self._descriptor_schema = None
-        #: The process used in this data object
-        self._process = None
-        #: ``Sample`` containing ``Data`` object
-        self._sample = None
 
         #: ``ResolweQuery`` containing parent ``Data`` objects (lazy loaded)
         self._parents = None
         #: ``ResolweQuery`` containing child ``Data`` objects (lazy loaded)
         self._children = None
 
-        #: checksum field calculated on inputs
-        self.checksum = None
-        #: indicate whether `descriptor` doesn't match `descriptor_schema` (is dirty)
-        self.descriptor_dirty = None
-        #: annotation data, with the form defined in descriptor_schema
-        self.descriptor = None
-        #: duplicated
-        self.duplicated = None
-        #: actual input values
-        self.input = None
-        #: process cores
-        self.process_cores = None
-        #: error log message (list of strings)
-        self.process_error = None
-        #: info log message (list of strings)
-        self.process_info = None
-        #: process memory
-        self.process_memory = None
-        #: process progress in percentage
-        self.process_progress = None
-        #: Process algorithm return code
-        self.process_rc = None
-        #: warning log message (list of strings)
-        self.process_warning = None
-        #: actual output values
-        self.output = None
-        #: process_resources
-        self.process_resources = None
-        #: size
-        self.size = None
-        #: scheduled
-        self.scheduled = None
         #: process status - Possible values:
         #: UP (Uploading - for upload processes),
         #: RE (Resolving - computing input data objects)
@@ -118,80 +98,22 @@ class Data(BaseResolweResource):
         #: OK (Done)
         #: ER (Error)
         #: DR (Dirty - Data is dirty)
-        self.status = None
-        #: data object's tags
-        self.tags = None
 
         super().__init__(resolwe, **model_data)
 
-    def update(self):
-        """Clear cache and update resource fields from the server."""
+    def _update_fields(
+        self, payload: dict[str, Any], data_source: DataSource = DataSource.USER
+    ):
+        """Update fields of the local resource based on the server values."""
+        # The payload contains collection information and sample information on the top
+        # level. The collection also contains the sample information, but only as id.
+        # Replace the id with the actual info.
+        if collection_payload := payload.get("collection"):
+            if sample_payload := payload.get("entity"):
+                sample_payload["collection"] = collection_payload
         self._children = None
-        self._collection = None
-        self._descriptor_schema = None
         self._parents = None
-        self._process = None
-        self._sample = None
-
-        super().update()
-
-    @property
-    def process(self):
-        """Get process."""
-        return self._process
-
-    @process.setter
-    def process(self, payload):
-        """Set process."""
-        self._resource_setter(payload, Process, "_process")
-
-    @property
-    def descriptor_schema(self):
-        """Get descriptor schema."""
-        return self._descriptor_schema
-
-    @descriptor_schema.setter
-    def descriptor_schema(self, payload):
-        """Set descriptor schema."""
-        self._resource_setter(payload, DescriptorSchema, "_descriptor_schema")
-
-    @property
-    def sample(self):
-        """Get sample."""
-        if self._sample is None and self._original_values.get("entity", None):
-            # The collection data is only serialized on the top level. Replace the
-            # data inside 'entity' with the actual collection data.
-            entity_values = self._original_values["entity"].copy()
-            entity_values["collection"] = self._original_values.get("collection", None)
-            self._sample = Sample(resolwe=self.resolwe, **entity_values)
-        return self._sample
-
-    @sample.setter
-    def sample(self, payload):
-        """Set sample."""
-        self._resource_setter(payload, Sample, "_sample")
-
-    @property
-    def collection(self):
-        """Get collection."""
-        return self._collection
-
-    @collection.setter
-    def collection(self, payload):
-        """Set collection."""
-        self._resource_setter(payload, Collection, "_collection")
-
-    @property
-    @assert_object_exists
-    def started(self):
-        """Get start time."""
-        return parse_resolwe_datetime(self._original_values["started"])
-
-    @property
-    @assert_object_exists
-    def finished(self):
-        """Get finish time."""
-        return parse_resolwe_datetime(self._original_values["finished"])
+        super()._update_fields(payload, data_source)
 
     @property
     @assert_object_exists
@@ -251,7 +173,12 @@ class Data(BaseResolweResource):
             {"resource_overrides": {self.id: overrides}}
         )
 
-    def _files_dirs(self, field_type, file_name=None, field_name=None):
+    def _files_dirs(
+        self,
+        field_type: str,
+        file_name: Optional[str] = None,
+        field_name: Optional[str] = None,
+    ) -> list[str]:
         """Get list of downloadable fields."""
         download_list = []
 
@@ -283,16 +210,20 @@ class Data(BaseResolweResource):
 
         return download_list
 
-    def _get_dir_files(self, dir_name):
-        files_list, dir_list = [], []
+    def _get_dir_files(self, dir_name: str) -> list[str]:
+        files_list: list[str] = []
+        dir_list: list[str] = []
 
         dir_url = urljoin(self.resolwe.url, "data/{}/{}".format(self.id, dir_name))
         if not dir_url.endswith("/"):
             dir_url += "/"
+        assert self.resolwe.session is not None
         response = self.resolwe.session.get(dir_url, auth=self.resolwe.auth)
         response = json.loads(response.content.decode("utf-8"))
 
+        assert isinstance(response, list), "Invalid response from server."
         for obj in response:
+            assert isinstance(obj, dict), "Invalid response from server."
             obj_path = "{}/{}".format(dir_name, obj["name"])
             if obj["type"] == "directory":
                 dir_list.append(obj_path)
@@ -306,17 +237,12 @@ class Data(BaseResolweResource):
         return files_list
 
     @assert_object_exists
-    def files(self, file_name=None, field_name=None):
+    def files(
+        self, file_name: Optional[str] = None, field_name: Optional[str] = None
+    ) -> list[str]:
         """Get list of downloadable file fields.
 
         Filter files by file name or output field.
-
-        :param file_name: name of file
-        :type file_name: string
-        :param field_name: output field name
-        :type field_name: string
-        :rtype: List of tuples (data_id, file_name, field_name, process_type)
-
         """
         file_list = self._files_dirs("file", file_name, field_name)
 
@@ -331,7 +257,7 @@ class Data(BaseResolweResource):
         field_name: Optional[str] = None,
         download_dir: Optional[str] = None,
         show_progress: bool = True,
-    ):
+    ) -> list[str]:
         """Download Data object's files and directories.
 
         Download files and directories from the Resolwe server to the
@@ -392,7 +318,7 @@ class Data(BaseResolweResource):
             destination_file_path,
         )
 
-    def stdout(self):
+    def stdout(self) -> str:
         """Return process standard output (stdout.txt file content).
 
         Fetch stdout.txt file from the corresponding Data object and return the
@@ -405,6 +331,7 @@ class Data(BaseResolweResource):
             raise ValueError("stdout.txt file is not available for workflows.")
         output = b""
         url = urljoin(self.resolwe.url, "data/{}/stdout.txt".format(self.id))
+        assert self.resolwe.session is not None
         response = self.resolwe.session.get(url, stream=True, auth=self.resolwe.auth)
         if not response.ok and self.status in ["UP", "RE", "WT", "PP", "DR"]:
             raise ValueError(
@@ -419,11 +346,13 @@ class Data(BaseResolweResource):
         return output.decode("utf-8")
 
     @assert_object_exists
-    def duplicate(self):
+    def duplicate(self) -> "Data":
         """Duplicate (make copy of) ``data`` object.
 
         :return: Duplicated data object
         """
         task_data = self.api().duplicate.post({"ids": [self.id]})
-        background_task = BackgroundTask(resolwe=self.resolwe, **task_data)
+        background_task = BackgroundTask(
+            resolwe=self.resolwe, initial_data_source=DataSource.SERVER, **task_data
+        )
         return self.resolwe.data.get(id__in=background_task.result())